@damned removed card_view.update_position from logical
had to generate a new event - is looking like an event similar
to moving/moved - but doesn't get raised externally because
this is already a move based on external event (e.g. card move
on another browser).

perhaps a more general scheme for this would be a flag or
source field for event